Bytecode is 16-bits. AMAZING MEMORY USE REDUCTION IN SANITY.
uintptr_t is 64 bits on my system. I assumed that the memory reduction would
be significant, because bytecode takes a lot of memory. But I thought that
signatures and values took up a significant portion of memory. I was wrong.

(Numbers reported by valgrind)
sanity.ly, before: 402,872 (402 KB).

After...

208,370 (208 KB).

It's...very close to half of the entire memory use. This is far above and
beyond what I expected in memory savings. I made sure that this did not break
mod_lily.

Wow. Just...wow.

+++ src/lily_cls_hash.c
@@ -1,7 +1,7 @@
 #include "lily_impl.h"
 #include "lily_vm.h"
 
-void lily_hash_get(lily_vm_state *vm, lily_function_val *self, uintptr_t *code)
+void lily_hash_get(lily_vm_state *vm, lily_function_val *self, uint16_t *code)
 {
     lily_value **vm_regs = vm->vm_regs;
     lily_value *input = vm_regs[code[0]];
